tests/wpt/tests/css/css-contain/content-visibility/content-visibility-091.html

<!doctype HTML>
<meta charset="utf8">
<title>CSS Content Visibility: getComputedStyle works in hidden c-v subtree</title>
<link rel="author" title="Rob Buis" href="mailto:rbuis@igalia.com">
<link rel="help" href="https://drafts.csswg.org/css-contain/#content-visibility">
<meta name="assert" content="getComputedStyle works in hidden c-v subtree">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
#container { content-visibility: hidden }
</style>

<div id=container style="width:200px">
<div id=target style="width:50%"></div>
</div>

<script>
test(() => {
  const computedStyle = getComputedStyle(target);
  assert_equals(computedStyle.width, "100px");
}, "getComputedStyle works in hidden c-v subtree");
</script>
